modularize test files
