+++ .gitignore
@@ -18,4 +18,5 @@ client/dist/vendors-node_modules_react-icons_go_index_esm_js.bundle.js
 client/dist/vendors-node_modules_react-icons_md_index_esm_js-node_modules_react-spring_dist_react-spring_esm_js.bundle.js
 client/dist/vendors-node_modules_react-icons_md_index_esm_js.bundle.js
 client/dist/vendors-node_modules_react-spring_dist_react-spring_esm_js.bundle.js
+client/dist/vendors-node_modules_react-icons_fi_index_esm_js.bundle.js
 package-lock.json
