+++ client/test/product-questions-tests/Questions.test.jsx
@@ -5,7 +5,7 @@
 import React from 'react';
 import { render, screen, waitFor } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import ProductQuestions from '../src/components/product-questions/ProductQuestions.jsx';
+import ProductQuestions from '../../src/components/product-questions/ProductQuestions.jsx';
 
 describe('Product Questions Component', () => {
   test('it displays list of questions', async () => {
